+++ hyvar.ui.test01/src/hyvar/ui/test01/GlgAnimationValue2.java
@@ -1,6 +1,5 @@
 
 package hyvar.ui.test01;
-import com.genlogic.*;
 
 //////////////////////////////////////////////////////////////////////////
 public class GlgAnimationValue2
@@ -62,7 +61,7 @@ public class GlgAnimationValue2
       
 
        case RANDOM_INT:
-         value = (double) ( (int) GlgObject.Rand( min, max ) );
+         value = ( (int) GlgObject.Rand( min, max ) );
          break;
 
        default:
